Plotly demo finished, soon in approval for merge
